requirements.txt
asgiref==3.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.4.5
django-filter==21.1
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.0.4
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2020.1
sqlparse==0.3.1
